launch_shinystan_demo {shinystan}	R Documentation

'ShinyStan' demo

Description

'ShinyStan' demo

Usage

launch_shinystan_demo(demo_name = "eight_schools",
  rstudio = getOption("shinystan.rstudio"), ...)

Arguments

demo_name
The name of the demo. Currently "eight_schools" is the only option, but additional demos may be available in future releases.

eight_schools

Hierarchical meta-analysis model. See Meta Analysis chapter of the 'Stan' manual http://mc-stan.org/users/documentation/.

rstudio
Only relevant for 'RStudio' users. The default (FALSE) is to launch the app in the user's default web browser rather than the pop-up Viewer provided by 'RStudio'. Users can change the default to TRUE by setting the global option options(shinystan.rstudio = TRUE).

...
Optional arguments passed to runApp.

Value

An S4 shinystan object.

See Also

launch_shinystan to launch the 'ShinyStan' interface using a particular shinystan object.

as.shinystan for creating shinystan objects.

Examples

## Not run: 
# launch demo but don't save a shinystan object
launch_shinystan_demo() 

# launch demo and save the shinystan object for the demo 
sso_demo <- launch_shinystan_demo()

## End(Not run)

[Package shinystan version 2.5.0 Index]