S4 shinystan objects

See as.shinystan for documentation on creating shinystan objects and eight_schools for an example object.

Slots

model_name

("character") Model name.

param_names

("character") Parameter names.

param_dims

("list") Parameter dimensions.

posterior_sample

("array") MCMC sample.

summary

("matrix") Summary stats for posterior_sample.

sampler_params

("list") Sampler parameters (for certain Stan models only).

n_chain

("integer") Number of chains.

n_iter

("integer") Number of iterations per chain.

n_warmup

("integer") Number of warmup iterations per chain.

user_model_info

("character") Notes to display on the Notepad page in the 'ShinyStan' GUI.

model_code

("character") Model code to display on the Model Code page in the 'ShinyStan' GUI.

misc

("list") Miscellaneous, for internal use.

See also

as.shinystan for creating shinystan objects. drop_parameters to remove parameters from a shinystan object. generate_quantity to add a new quantity to a shinystan object. shinystan-metadata to view or change metadata associated with a shinystan object.